Modify details gathered from last meeting. Add policies to identify if user can access stuff. Products and its formats

+++ admin_page/app/Http/Controllers/OrderController.php
@@ -10,6 +10,7 @@ use App\Client;
 use App\Exports\OrdersExport;
 use App\Product;
 use App\TimeBlock;
+use App\ProductFormat;
 
 class OrderController extends Controller
 {
@@ -20,6 +21,8 @@ class OrderController extends Controller
      */
     public function index(FilterRequest $request)
     {
+        $this->authorize('view', Order::class);
+
         $client_type = $request->query('client_type');
         $time_interval_start = $request->query('time_interval_start');
         $time_interval_end = $request->query('time_interval_end');
@@ -56,13 +59,16 @@ class OrderController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Order::class);
+
         $towns = Town::pluck('name', 'id');
         $products = Product::pluck('name', 'id');
+        $formats = ProductFormat::where('product_id', Product::compounded()->first()->id)->pluck('name', 'id');
         $deliveryStatuses = Order::getDeliveryStatuses();
         $paymentStatuses = Order::getPaymentStatuses();
         $timeBlocks = TimeBlock::orderBy('start')->get();
 
-        return view('orders.create', compact('towns', 'products', 'deliveryStatuses', 'paymentStatuses', 'timeBlocks'));
+        return view('orders.create', compact('towns', 'products', 'deliveryStatuses', 'paymentStatuses', 'timeBlocks', 'formats'));
     }
 
     /**
@@ -73,6 +79,8 @@ class OrderController extends Controller
      */
     public function store(OrderRequest $request)
     {
+        $this->authorize('create', Order::class);
+
         Order::createFromForm($request->validated());
 
         return redirect()->route('orders.index')
@@ -87,6 +95,8 @@ class OrderController extends Controller
      */
     public function show(Order $order)
     {
+        $this->authorize('view', Order::class);
+
         return view('orders.show', compact('order'));
     }
 
@@ -98,6 +108,8 @@ class OrderController extends Controller
      */
     public function edit(Order $order)
     {
+        $this->authorize('update', Order::class);
+
         return view('orders.edit', compact('order'));
     }
 
@@ -110,6 +122,8 @@ class OrderController extends Controller
      */
     public function update(OrderRequest $request, Order $order)
     {
+        $this->authorize('update', Order::class);
+
         $order->update($request->validated());
 
         return redirect()->route('orders.index')
@@ -124,6 +138,8 @@ class OrderController extends Controller
      */
     public function destroy(Order $order)
     {
+        $this->authorize('delete', Order::class);
+
         $order->delete();
 
         return redirect()->route('orders.index', [
@@ -138,6 +154,8 @@ class OrderController extends Controller
 
     public function delivered(Order $order)
     {
+        $this->authorize('deliver', Order::class);
+
         $order->delivered();
 
         return redirect()->route('orders.index', [
@@ -152,6 +170,8 @@ class OrderController extends Controller
 
     public function paid(Order $order)
     {
+        $this->authorize('payment', Order::class);
+
         $order->paid();
 
         return redirect()->route('orders.index', [
